feat(): add default listener functionality
Added new static and instance setDefaultListener methods
to allow signal creators to set default listeners to be used
when no other listener is added. This should typically be used
to report errors or log when no listener is added as it can
be considered a programmer error in some use cases.